services/describe-image/requirements.txt
accelerate==1.9.0
fastapi==0.116.1
httpx==0.24.0
numpy==1.26.4
openai==1.97.1
Pillow==11.0.0
python-dotenv==1.1.1
qwen-vl-utils==0.0.11
requests==2.32.4
urllib3==2.5.0
uvicorn[standard]==0.35.0
# Transformers for local Qwen vision model
torch==2.7.1+cu128 --index-url https://download.pytorch.org/whl/cu128
torchvision==0.22.1+cu128 --index-url https://download.pytorch.org/whl/cu128
transformers==4.54.1
accelerate==1.9.0
